refactor: move bulk logic to kv module
