+++ src/commands/kv/bulk/mod.rs
@@ -1,7 +1,5 @@
 pub mod delete;
 pub mod put;
 
-pub use delete::delete;
-pub use put::put;
-
-const MAX_PAIRS: usize = 10000;
+pub use delete::run as delete;
+pub use put::run as put;
