+++ src/kv/mod.rs
@@ -1,1 +1,2 @@
+pub mod bulk;
 pub mod namespace;
